requirements.txt
tensorflow==2.0.0b1
numpy==1.17.4
Flask==1.1.1
gunicorn==20.0.4
Pillow